% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_binary_clustering.R
\name{cluster_chem_data}
\alias{cluster_chem_data}
\title{Run logistic singular value decomposition on the binary molecular features}
\usage{
cluster_chem_data(receptor_type, data, rank = 3, filter_type = "Pubchem")
}
\arguments{
\item{receptor_type}{Used for the kmeans plot from the SVD results, this is a characer string such as "Androgen"}

\item{data}{matrix or dataframe that has all binary data except for possibly Name, Label_names, and Label, which are removed}

\item{rank}{rank of the SVD that decomposes the binary molecular feature data before using the PCs in kmeans.}
}
\value{
Results from the logistic SVD

Plot of Kmeans for visual inspection before creating clusters
}
\description{
This function takes in a binary matrix, assuming that the variable names include the fingerprint name such as "Pubchem"
}
