% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_load_aggregate_data.R
\name{load_chem_data}
\alias{load_chem_data}
\title{Load Chemical Fingerprint Data}
\usage{
load_chem_data(path, file_type)
}
\arguments{
\item{path}{Path to the input file}

\item{file_type}{type of file (xlsx, csv, or txt)}
}
\value{
A matrix
}
\description{
This function loads a file as a matrix. All explanatory data should be binary,
indicating presence or absence of a bit structure, the first column
should have the chemical name in order to join on additional physical data late.
}
