% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_plot_logistic_SVD_kmeans.R
\name{plot_SVD_kmeans}
\alias{plot_SVD_kmeans}
\title{Plot the first two PCs from logistic SVD for molecular features and color by k-means results}
\usage{
plot_SVD_kmeans(
  receptor_type,
  log_SVD_results,
  binary_data,
  k = 4,
  endo_ligand = FALSE,
  endo_ligand_id,
  additional_label = TRUE,
  outcome_label = "ASSAY_OUTCOME"
)
}
\arguments{
\item{receptor_type}{type of receptor for labeling plots}

\item{log_SVD_results}{Results from the SVD to get data from for plotting}

\item{k}{number of groups from kmeans from after viewing elbow plot or other methods of determining groups}
}
\value{
Plot of SVD results colored by kmeans groups
}
\description{
This function takes in a binary matrix, assuming that the variable names include the fingerprint name such as "Pubchem"
}
