% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_run_MCMC_logic.R
\name{run_logic_MCMC}
\alias{run_logic_MCMC}
\title{Wrapper for MCMC logic regression package to run monte carlo fitting of logic trees on each labeled subgroup for those found to be active binders}
\usage{
run_logic_MCMC(
  outcome_label,
  filter_type,
  binary_data,
  subgroups,
  subgroup_label,
  mx_size,
  ntrees,
  n_burn_in,
  niter,
  hyperparameters = log(2),
  outcome_permuted = FALSE,
  top_fraction = 0.2
)
}
\arguments{
\item{log_SVD_results}{Results from the SVD to get data from for plotting}

\item{receptor_type}{type of receptor for labeling plots}

\item{k}{number of groups from kmeans from after viewing elbow plot or other methods of determining groups}
}
\value{
Plot of SVD results colored by kmeans groups
}
\description{
This function takes in a binary matrix, assuming that the variable names include the fingerprint name such as "Pubchem"
}
